Allow user to enable/disable graph with positional argument

+++ config.py
@@ -6,7 +6,7 @@ SIMULATION_ENDED = False
 SIMULATION_STARTED = False
 GRAPH_ENABLED = True
 
-def setup_graph_and_canvas(fastGraph):
+def setup_graph_and_canvas(fastGraph, enable_graph):
 
     global GRAPH
     global CANVAS
@@ -14,11 +14,12 @@ def setup_graph_and_canvas(fastGraph):
     global NUCLEUS
     global ELECTRON
 
-    GRAPH = vpython.graph(title='Electron Energy graph', xtitle='Time', ytitle='Energy', align='right', width=500, fast=fastGraph)
+    if enable_graph:
+        GRAPH = vpython.graph(title='Electron Energy graph', xtitle='Time', ytitle='Energy', align='right', width=500, fast=fastGraph)
 
-    # Create a gvbars object to align GRAPH on top right
-    bar = vpython.gvbars()
-    bar.plot(0, 0)
+        # Create a gvbars object to align GRAPH on top right
+        bar = vpython.gvbars()
+        bar.plot(0, 0)
     
     CANVAS = vpython.canvas(visible=True)
     CANVAS.select()
@@ -31,4 +32,5 @@ def setup_graph_and_canvas(fastGraph):
     ELECTRON = vpython.sphere(radius=1,
                               color=vpython.color.yellow,
                               make_trail=True)
-    bar.delete()
+    if enable_graph:
+        bar.delete()
